+++ .github/workflows/e2e.yml
@@ -44,8 +44,8 @@ jobs:
             class: org.apache.dolphinscheduler.e2e.cases.ProjectE2ETest
           - name: Workflow
             class: org.apache.dolphinscheduler.e2e.cases.WorkflowE2ETest
-          - name: Resource
-            class: org.apache.dolphinscheduler.e2e.cases.ResourceE2ETest
+          - name: FileManage
+            class: org.apache.dolphinscheduler.e2e.cases.FileManageE2ETest
     env:
       RECORDING_PATH: /tmp/recording-${{ matrix.case.name }}
     steps:
